+++ index.html
@@ -37,6 +37,6 @@
     <div id="scoreboard" style="position:absolute; top:10%; left:50%; transform:translateX(-50%); padding:20px; display:none;">
       <table id="score-table"></table>
     </div>
-  <script src="dist/main.js"></script>
+  <script type="module" src="dist/main.js"></script>
 </body>
 </html>
