+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "module": "none",
+    "module": "es6",
     "outDir": "dist",
     "strict": true
   },
